requirements.txt
lightning==2.0.8
matplotlib==3.7.2
scipy==1.11.2
pandas==2.1.0
numpy==1.25.2
click==8.1.7
plotly==5.16.1
wandb==0.15.9
class_resolver==0.4.2
atlasify==0.8.0
--find-links https://data.pyg.org/whl/torch-2.0.0+cu117.html
torch-cluster==1.6.1
pytest==7.4.2
